requirements.txt
dash
dash-bootstrap-components
pandas
sqlalchemy
psycopg2-binary
python-dotenv 